commands relocation
