+++ command/root/root.go
@@ -14,13 +14,13 @@ import (
 	"github.com/0xPolygon/polygon-edge/command/mint"
 	"github.com/0xPolygon/polygon-edge/command/monitor"
 	"github.com/0xPolygon/polygon-edge/command/peers"
-	"github.com/0xPolygon/polygon-edge/command/polybft"
 	"github.com/0xPolygon/polygon-edge/command/regenesis"
 	"github.com/0xPolygon/polygon-edge/command/secrets"
 	polybftsecrets "github.com/0xPolygon/polygon-edge/command/secrets/init"
 	"github.com/0xPolygon/polygon-edge/command/server"
 	"github.com/0xPolygon/polygon-edge/command/status"
 	"github.com/0xPolygon/polygon-edge/command/txpool"
+	"github.com/0xPolygon/polygon-edge/command/validator"
 	"github.com/0xPolygon/polygon-edge/command/version"
 )
 
@@ -55,10 +55,10 @@ func (rc *RootCommand) registerSubCommands() {
 		server.GetCommand(),
 		license.GetCommand(),
 		polybftsecrets.GetCommand(),
-		polybft.GetCommand(),
 		bridge.GetCommand(),
 		regenesis.GetCommand(),
 		mint.GetCommand(),
+		validator.GetCommand(),
 	)
 }
